+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/template
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.9
+
 ## 0.2.9
 
 ### Patch Changes